+++ src/main/java/com/pms/document/service/IDocumentDetailsService.java
@@ -13,10 +13,10 @@ import com.pms.document.entity.DocumentDetails;
 public interface IDocumentDetailsService {
 	
 	public List<DocumentDetails> getDocumentDetails();
-	public <Optional>DocumentDetails getDocumentDetail(int documentDetailId);
+	public <Optional>DocumentDetails getDocumentDetail(Long documentDetailId);
 	public DocumentDetails createDocumentDetails(DocumentDetails documentDetails);
-	public DocumentDetails updateDocumentDetails(int documentDetailsId, DocumentDetails documentDetails);
-	public boolean deleteDocumentDetails(int documentDetailsId);
+	public DocumentDetails updateDocumentDetails(Long documentDetailsId, DocumentDetails documentDetails);
+	public boolean deleteDocumentDetails(Long documentDetailsId);
 	 public List<DocumentDetails> search(String documentTypeEnum, String documentNumber);
 
 }
