+++ src/main/java/com/pms/document/service/IDocumentTypeService.java
@@ -23,7 +23,7 @@ static final Logger logger = LoggerFactory.getLogger(IDocumentTypeService.class)
 	public List<DocumentType> getDocumentTypes();
 	public DocumentType getDocumentType(Long documentTypeId);
 	public DocumentType createDocumentType(DocumentType documentType);
-	public DocumentType updateDocumentType(int documentTypeId,DocumentType documentType);
+	public DocumentType updateDocumentType(Long documentTypeId,DocumentType documentType);
 	public boolean deleteDocumentType(Long documentTypeId);
 	public DocumentType findByIdAndHotelId(Long id);
 	public List<DocumentType> search(String shortName, String documentTypeName,String documentTypeDescription);
