+++ src/main/java/com/pms/guestdetails/service/IGuestDetailsService.java
@@ -13,9 +13,9 @@ import com.pms.guestdetails.GuestDetails;
 public interface IGuestDetailsService {
 	
 	public List<GuestDetails> getGuestDetails();
-	public GuestDetails getGuestDetail(int guestDetailsId);
+	public GuestDetails getGuestDetail(Long guestDetailsId);
 	public GuestDetails createGuestDetail(GuestDetails guestDetails);
-	public GuestDetails updateGuestDetails(int guestDetailsId,GuestDetails guestDetails);
+	public GuestDetails updateGuestDetails(Long guestDetailsId,GuestDetails guestDetails);
 	public boolean deleteSoftStayDetails(Integer id);
